Rename fanoutprocessor to FanOutConnector
This implements https://github.com/open-telemetry/opentelemetry-service/issues/283
